Update and rename creating_a_turtle.rs to draw_turtle.rs
